Time: 189 ms (20.95%), Space: 36.8 MB (74.47%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -28,7 +28,7 @@ public:
             
         }
         
-        return res%(int)1000000007;
+        return res%1000000007;
         
     }
 };
